src/requirements.txt
google-api-core
google-cloud-storage
